requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
aiohttp>=3.9.3
openai>=1.12.0
docker>=7.0.0
paramiko>=3.4.0
mcp>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.3
ntplib>=0.4.0
google-generativeai>=0.8.3
google-genai>=1.0.0
chromadb>=0.4.24
